Renamed 'events' package to 'legacy-events'
