+++ packages/legacy-events/EventBatching.js
