+++ packages/legacy-events/ReactGenericBatching.js
