+++ packages/legacy-events/ReactSyntheticEventType.js
