+++ packages/legacy-events/SyntheticEvent.js
