+++ packages/legacy-events/__tests__/accumulateInto-test.internal.js
@@ -13,7 +13,7 @@ let accumulateInto;
 
 describe('accumulateInto', () => {
   beforeEach(() => {
-    accumulateInto = require('events/accumulateInto').default;
+    accumulateInto = require('legacy-events/accumulateInto').default;
   });
 
   it('throws if the second item is null', () => {
