+++ packages/legacy-events/accumulate.js
