+++ packages/legacy-events/accumulateInto.js
