+++ packages/legacy-events/forEachAccumulated.js
