+++ packages/legacy-events/package.json
@@ -1,5 +1,5 @@
 {
   "private": true,
-  "name": "events",
+  "name": "legacy-events",
   "version": "0.0.0"
 }
